requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
blessings==1.7
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
commonmark==0.9.1
curtsies==0.3.10
cwcwidth==0.1.6
enquiries==0.1.0
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
Pygments==2.11.2
PySocks==1.7.1
python-dotenv==0.19.2
requests==2.27.1
rich==11.1.0
six==1.16.0
soupsieve==2.3.1
tqdm==4.62.3
urllib3==1.26.8
yarl==1.7.2
zippyshare-downloader==0.3.0
